Initial Submittal
Submittal of code from Version 1. Version 1 was based on using two AHRS to measure the mast rotation and then correcting the NMEA2000 message sent to the display.  This method was proven to be inaccurate and too finicky with calibration of the AHRS and the sensor method is now being changed to an SPS-A100D-HAWS from Honeywell.  Ensuing code changes will be related to that integration.